+++ src/test/groovy/ru/vyarus/dropwizard/guice/test/jupiter/debug/SetupObjectsLogTest.java
@@ -1,4 +1,4 @@
-package ru.vyarus.dropwizard.guice.test.jupiter.setup;
+package ru.vyarus.dropwizard.guice.test.jupiter.debug;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.platform.testkit.engine.EngineTestKit;
 import ru.vyarus.dropwizard.guice.support.AutoScanApplication;
+import ru.vyarus.dropwizard.guice.test.TestSupport;
 import ru.vyarus.dropwizard.guice.test.jupiter.TestGuiceyApp;
 import ru.vyarus.dropwizard.guice.test.jupiter.env.EnableSetup;
 import ru.vyarus.dropwizard.guice.test.jupiter.env.TestEnvironmentSetup;
@@ -30,6 +31,7 @@ public class SetupObjectsLogTest {
 
     @Test
     void checkSetupOutputForAnnotation() {
+        TestSupport.debugExtensions();
         EngineTestKit
                 .engine("junit-jupiter")
                 .configurationParameter("junit.jupiter.conditions.deactivate", "org.junit.*DisabledCondition")
@@ -44,19 +46,20 @@ public class SetupObjectsLogTest {
 
         Assertions.assertTrue(output.replaceAll("\\$\\$Lambda\\$\\d+/\\d+(x[a-z\\d]+)?", "\\$\\$Lambda\\$111/1111111")
                 .replaceAll("\\) {8,}\t", ")        \t")
-                .contains("Guicey test extensions:\n" +
+                .contains("Guicey test extensions (Test1.):\n" +
                         "\n" +
                         "\tSetup objects = \n" +
-                        "\t\tExt1                         (r.v.d.g.t.j.s.SetupObjectsLogTest)        \t@TestGuiceyApp\n" +
-                        "\t\tExt2                         (r.v.d.g.t.j.s.SetupObjectsLogTest)        \t@TestGuiceyApp\n" +
-                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Base.base1\n" +
-                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Base.base2\n" +
-                        "\t\tSetupObjectsLogTest$Test1$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Test1.ext1\n" +
-                        "\t\tSetupObjectsLogTest$Test1$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Test1.ext2\n"));
+                        "\t\tExt1                         (r.v.d.g.t.j.d.SetupObjectsLogTest)        \t@TestGuiceyApp\n" +
+                        "\t\tExt2                         (r.v.d.g.t.j.d.SetupObjectsLogTest)        \t@TestGuiceyApp\n" +
+                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Base.base1\n" +
+                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Base.base2\n" +
+                        "\t\tSetupObjectsLogTest$Test1$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Test1.ext1\n" +
+                        "\t\tSetupObjectsLogTest$Test1$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Test1.ext2\n"));
     }
 
     @Test
     void checkSetupOutputForManualRegistration() {
+        TestSupport.debugExtensions();
         EngineTestKit
                 .engine("junit-jupiter")
                 .configurationParameter("junit.jupiter.conditions.deactivate", "org.junit.*DisabledCondition")
@@ -71,17 +74,17 @@ public class SetupObjectsLogTest {
 
         Assertions.assertTrue(output.replaceAll("\\$\\$Lambda\\$\\d+/\\d+(x[a-z\\d]+)?", "\\$\\$Lambda\\$111/1111111")
                 .replaceAll("\\) {8,}\t", ")        \t")
-                .contains("Guicey test extensions:\n" +
+                .contains("Guicey test extensions (Test2.):\n" +
                         "\n" +
                         "\tSetup objects = \n" +
-                        "\t\tExt1                         (r.v.d.g.t.j.s.SetupObjectsLogTest)        \t@RegisterExtension class\n" +
-                        "\t\tExt2                         (r.v.d.g.t.j.s.SetupObjectsLogTest)        \t@RegisterExtension class\n" +
-                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@RegisterExtension instance\n" +
-                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@RegisterExtension instance\n" +
-                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Base.base1\n" +
-                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Base.base2\n" +
-                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Test2.ext1\n" +
-                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.setup)        \t@EnableSetup field Test2.ext2\n"));
+                        "\t\tExt1                         (r.v.d.g.t.j.d.SetupObjectsLogTest)        \t@RegisterExtension class\n" +
+                        "\t\tExt2                         (r.v.d.g.t.j.d.SetupObjectsLogTest)        \t@RegisterExtension class\n" +
+                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@RegisterExtension instance\n" +
+                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@RegisterExtension instance\n" +
+                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Base.base1\n" +
+                        "\t\tSetupObjectsLogTest$Base$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Base.base2\n" +
+                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Test2.ext1\n" +
+                        "\t\tSetupObjectsLogTest$Test2$$Lambda$111/1111111 (r.v.d.g.t.j.debug)        \t@EnableSetup field Test2.ext2\n"));
     }
 
     public static class Base {